api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
pydantic==2.9.2
python-docx==1.1.2
pypdf==4.3.1
google-cloud-storage==2.18.2
python-multipart==0.0.9